With JPA annotation

+++ DeliveryService/src/main/java/anna/deliveryservice/repository/CustomerRepository.java
@@ -13,5 +13,5 @@ import anna.deliveryservice.domain.Customer;
  */
 public interface CustomerRepository {
     Customer update(Customer customer);
-    Customer findById(int id);
+    Customer findById(long id);
 }
